% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_dense.R
\name{BDense}
\alias{BDense}
\title{Create a Dense layer that has a standard normal prior for weights and biases}
\usage{
BDense(
  in_size,
  out_size,
  activation = c("identity", "sigmoid", "tanh", "relu")
)
}
\arguments{
\item{in_size}{input size}

\item{out_size}{output size}

\item{activation}{activation function}
}
\description{
Create a Dense layer that has a standard normal prior for weights and biases
}
