% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{BNN}
\alias{BNN}
\title{Create the actual BNN model}
\usage{
BNN(net, y, x)
}
\arguments{
\item{net}{A BNN network created usign \code{\link{make_net}}}

\item{y}{a numerical vector of length N}

\item{x}{a numerical matrix with each column being one input and one row being one variable
For example, for a network with one input, N = 100 observations, nrow(x) = 1, ncol(x) = N}
}
\description{
This function follows after \code{\link{make_net}} and supplies the
BNN with the data and precompiles it
}
\examples{
y <- arima.sim(list(ar = 0.5), n = 500)
x <- matrix(y[1:499], nrow = 1)
y <- y[2:500]
quickbnnr_setup()
net <- Chain(DenseBNN(1, 1, "sigmoid"), DenseBNN(1, 1))
net <- make_net(net)
model <- BNN(net, y, x)
chains <- estimate(model, niter = 100, nchains = 4)

# library(bayesplot)
# mcmc_intervals(chains$draws)

}
