% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_recurrent.R
\name{BRNN}
\alias{BRNN}
\title{Adds a Baysian RNN layer to the network}
\usage{
BRNN(in_size, out_size, activation = c("tanh", "sigmoid", "identity", "relu"))
}
\arguments{
\item{in_size}{input size of the layer}

\item{out_size}{output size of the layer (i.e. the size
of the hidden state)}

\item{activation}{activation function; Standard is tanh or sigmoid}
}
\description{
Adds a Baysian RNN layer to the network
}
