% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.R
\name{Chain}
\alias{Chain}
\title{Create a network by chaining layers}
\usage{
Chain(...)
}
\arguments{
\item{...}{Layers separated by a comma}
}
\value{
Returns the julia specification of the network as a string
to be passed on to \code{\link{make_net}}
}
\description{
Create a network by chaining layers
}
\examples{
net <- Chain(DenseForcePosFirstWeight(1, 1, "sigmoid"), DenseBNN(1, 1))

}
