% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{clean_parameters}
\alias{clean_parameters}
\title{Clean parameters for better understanding}
\usage{
clean_parameters(est)
}
\arguments{
\item{est}{estimated model using \code{\link{estimate}}}
}
\description{
Some of the layers implemented take cumsums of variables to obtain
the actual NN parameters used. \code{\link{estimate}} will return
the parameters before taking the cumsum and thus can be difficult
to interpret. This method fixes this.
}
