% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_dense.R
\name{DenseForcePosFirstWeight}
\alias{DenseForcePosFirstWeight}
\title{Create a layer that enforces positive weights in the first column of the weight matrix.
This often solves some identification issues.}
\usage{
DenseForcePosFirstWeight(
  in_size,
  out_size,
  activation = c("identity", "sigmoid", "tanh", "relu")
)
}
\arguments{
\item{in_size}{input size}

\item{out_size}{output size}

\item{activation}{activation function}
}
\description{
Priors for all biases and weights not in the first weight matrix column are standard normal.
Priors for the weights in the first column are truncted standard normal (0, inf).
}
