% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.R
\name{DenseOrderedBias}
\alias{DenseOrderedBias}
\title{Create a Dense layer with ordered biases}
\usage{
DenseOrderedBias(
  in_size,
  out_size,
  activation = c("identity", "sigmoid", "tanh", "relu")
)
}
\arguments{
\item{in_size}{input size}

\item{out_size}{output size}

\item{activation}{activation function}
}
\description{
Priors for weights are standard normal. The ordering of biases is
implemented by putting a standard normal prior on the first bias,
and a truncated standard normal prior (0, inf) on the other biases.
The actual biases are then the cumulative sum of the biases.
}
