% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.R
\name{DenseOrderedWeights}
\alias{DenseOrderedWeights}
\title{Create a dense layer with ordered first weight column}
\usage{
DenseOrderedWeights(
  in_size,
  out_size,
  activation = c("identity", "sigmoid", "tanh", "relu")
)
}
\arguments{
\item{in_size}{input size}

\item{out_size}{output size}

\item{activation}{activation function}
}
\description{
Priors for biases and all weights in columns not being the first are
standard normal. The first column of the weight matrix is specified
by putting a standard normal prior on W(1, 1) and a truncated standard normal(0, inf)
on the remaining weights in that column. The weights in the column are then the cumsum
of those prior weights
}
