% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{estimate}
\alias{estimate}
\title{Estimating a BNN model using the NUTS default sampler in Turing}
\usage{
estimate(model, niter = 100, nchains = 1)
}
\arguments{
\item{model}{model formed using \code{\link{BNN}}}

\item{niter}{Number of MCMC iterations}

\item{nchains}{Number of Chains}
}
\value{
Returns a quickbnnr.estimate object with the number of iterations run, the number of chains,
the draws in a format that it can be used with bayesplot, and some additional internal information
}
\description{
Estimating a BNN model using the NUTS default sampler in Turing
}
