% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_random_symbol}
\alias{get_random_symbol}
\title{Creates a random string that is used as variable in julia}
\usage{
get_random_symbol()
}
\description{
Creates a random string that is used as variable in julia
}
