% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{make_net}
\alias{make_net}
\title{Create a BNN based on a specification}
\usage{
make_net(specification)
}
\arguments{
\item{specification}{QuickBNN.jl ChainBNN specification; Can be created using \code{\link{Chain}}}
}
\description{
Specification must be a valid ChainBNN description from the
QuickBNN.jl package. This can be created using the \code{\link{Chain}}
function and any of the provided layer types, such as \code{\link{DenseBNN}}
}
\examples{
y <- arima.sim(list(ar = 0.5), n = 500)
x <- matrix(y[1:499], nrow = 1)
y <- y[2:500]
quickbnnr_setup()
net <- Chain(DenseBNN(1, 1, "sigmoid"), DenseBNN(1, 1))
net <- make_net(net)
model <- BNN(net, y, x)
chains <- estimate(model, niter = 100, nchains = 4)

# library(bayesplot)
# mcmc_intervals(chains$draws)

}
