% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{quickbnnr_seed}
\alias{quickbnnr_seed}
\title{Sets a seed for replication purposes}
\usage{
quickbnnr_seed(seed)
}
\arguments{
\item{seed}{seed to be used}
}
\description{
Sets a seed for replication purposes
}
