% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{summary.quickbnnr.estimate}
\alias{summary.quickbnnr.estimate}
\title{Returns summary statistics of an estimated model}
\usage{
\method{summary}{quickbnnr.estimate}(est)
}
\arguments{
\item{est}{model estimated using \code{\link{estimate}}}
}
\description{
Returns summary statistics of an estimated model
}
