% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickbnnr.R
\name{quickbnnr_setup}
\alias{quickbnnr_setup}
\title{Set up of the Julia environment needed for QuickBNN}
\usage{
quickbnnr_setup(pkg_check = TRUE, nthreads = 4, ...)
}
\arguments{
\item{pkg_check}{Booloan; Whether to check if packages are isntalled; Default is TRUE}

\item{nthreads}{Int; Number of threads to use; Default is 4;}

\item{...}{addtitional params passed to JuliaCall::julia_setup}
}
\value{
Nothing
}
\description{
Set up of the Julia environment needed for QuickBNN
}
